Start using material design (starting from Hello World)
